chore: reorganize folder structure for better code organization
- Created dedicated pages/ folder for route components
- Moved Login, Dashboard, PropertyListingsPage, PropertyDetail to pages/
- Created context/ folder for React context providers
- Moved AuthContext from components/ to context/
- Updated all import paths across the application
- Improved separation of concerns between UI components and page logic
- Enhanced project maintainability and scalability
