+++ src/pages/PropertyDetail/PropertyDetail.module.scss
